# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LINKa Identity API
  version: 0.1.0
  description: >-
    Internal API for optional accounts, anonymous installations, encrypted email
    identities, organizations, consent, telemetry preference, privacy requests,
    and product-scoped tokens. Workload credentials are role- and product-scoped.
    Raw email and root UUIDs never appear in a response or token.
servers:
  - url: http://localhost:8080
security:
  - internalBearer: []
paths:
...
  /v1/consents:
    post:
      operationId: createConsent
      summary: Append an explicit versioned consent status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentRequest"
      responses:
        "201":
          description: Consent evidence recorded
          content:
            application/json:
              schema:
                type: object
                required: [id, recorded_at]
                properties:
                  id:
                    $ref: "#/components/schemas/UUID"
                  recorded_at:
                    type: string
                    format: date-time
  /v1/telemetry-preferences:
    put:
      operationId: setTelemetryPreference
      summary: Record explicit telemetry preference
      description: Client recorded_at is bounded and retained only as audit time. A transition to denied emits a suppression event with server time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryPreferenceRequest"
      responses:
        "200":
          description: Preference stored
          content:
            application/json:
              schema:
                type: object
                required: [preference, recorded_at]
                properties:
                  preference:
                    type: string
                    enum: [allowed, denied]
                  recorded_at:
                    type: string
                    format: date-time
  /v1/privacy-requests:
    post:
      operationId: createPrivacyRequest
      summary: Request deletion
      description: Anonymous installations can request product scope only.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivacyRequest"
      responses:
        "202":
          description: Request accepted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedStatus"
...
  /v1/privacy-requests/{id}:
    get:
      operationId: getPrivacyRequest
      summary: Read privacy-request status
      parameters:
        - $ref: "#/components/parameters/UUIDPath"
      responses:
        "200":
          description: Current request status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyRequestStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/internal/privacy-requests/{id}/status:
    post:
      operationId: updatePrivacyRequestStatus
      summary: Apply an audited privacy-request state transition
      parameters:
        - $ref: "#/components/parameters/UUIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivacyStatusUpdate"
      responses:
        "200":
          description: State transition applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  securitySchemes:
    internalBearer:
      type: http
      scheme: bearer
      description: Per-workload credential. Server-side roles and product scopes authorize each route.
  parameters:
    UUIDPath:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
  responses:
    BadRequest:
      description: Request failed strict validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Existing resource conflicts with the supplied data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    UUID:
      type: string
      format: uuid
    ProductID:
      type: string
      pattern: "^[a-z][a-z0-9_-]{1,63}$"
    OpaqueKey:
      type: string
      pattern: "^[a-f0-9]{64}$"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CreatedStatus:
      type: object
      required: [id, status]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        status:
          type: string
    SubjectFields:
      type: object
      required: [subject_type, subject_key]
      properties:
        subject_type:
          type: string
          enum: [person, account, installation]
        subject_key:
          $ref: "#/components/schemas/OpaqueKey"
...
    ConsentRequest:
      allOf:
        - $ref: "#/components/schemas/SubjectFields"
        - type: object
          required: [product_id, consent_type, policy_version, status, recorded_at]
          properties:
            product_id:
              $ref: "#/components/schemas/ProductID"
            consent_type:
              type: string
              minLength: 1
              maxLength: 100
            policy_version:
              type: string
              minLength: 1
              maxLength: 100
            status:
              type: string
              enum: [granted, withdrawn]
            recorded_at:
              type: string
              format: date-time
      unevaluatedProperties: false
    TelemetryPreferenceRequest:
      allOf:
        - $ref: "#/components/schemas/SubjectFields"
        - type: object
          required: [product_id, preference, recorded_at]
          properties:
            product_id:
              $ref: "#/components/schemas/ProductID"
            preference:
              type: string
              enum: [allowed, denied]
            recorded_at:
              type: string
              format: date-time
      unevaluatedProperties: false
    PrivacyRequest:
      allOf:
        - $ref: "#/components/schemas/SubjectFields"
        - type: object
          required: [request_type, scope]
          properties:
            request_type:
              type: string
              const: deletion
            scope:
              type: string
              enum: [product, all]
            product_id:
              $ref: "#/components/schemas/ProductID"
            anchor_product_id:
              $ref: "#/components/schemas/ProductID"
      unevaluatedProperties: false
    PrivacyRequestStatus:
      type: object
      required: [id, request_type, scope, status, requested_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        request_type:
          type: string
          enum: [export, deletion]
        scope:
          type: string
          enum: [product, all]
        product_id:
          $ref: "#/components/schemas/ProductID"
        status:
          type: string
          enum: [requested, processing, completed, rejected, cancelled]
        requested_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    PrivacyStatusUpdate:
      type: object
      additionalProperties: false
      required: [status, audit_note]
      properties:
        status:
          type: string
          enum: [processing, rejected, cancelled]
        audit_note:
          type: string
          minLength: 1
          maxLength: 1000